+++ eagle-stores/src/components/index.js
@@ -1,0 +1,4 @@
+export {Carousel} from "./Carousel/Carousel"
+export {Footer} from "./Footer/Footer"
+export {HomeCategory} from "./HomeCategory/HomeCategory"
+export {NavBar} from "./NavBar/NavBar"
