+++ eagle-stores/src/reducer/index.js
@@ -1,0 +1,1 @@
+export { DataReducer , initialistate } from "./DataReducer"
