+++ eagle-stores/src/utils/filterFunction.js
@@ -34,8 +34,7 @@ const sortByPrice = (data, sortBy) => {
   if (sortBy === "") return data;
   if (sortBy === "HighToLow") {
     return [...data].sort((a, b) => b.price - a.price);
-  } else if (sortBy === "LowToHigh")
-    return [...data].sort((a, b) => a.price - b.price);
+  } return [...data].sort((a, b) => a.price - b.price);
 };
 
 const priceRangeFilter = (data, maxValue) => {
